refactor(navigation): implement client and worker feature scaffolds
Add separate scaffolding for client and worker features to enhance
navigation and UI structure. This change introduces new components
for client and worker bottom bars and menus, improving code
organization and maintainability. The new structure allows for
better handling of navigation states and user interactions.
