+++ composeApp/src/commonMain/kotlin/com/example/seviya/app/App.kt
@@ -2,14 +2,6 @@ package com.example.seviya.app
 
 
 //esto se agrego para que sirva drante la migracion de booking
-import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.animation.core.tween
-import androidx.compose.animation.fadeIn
-import androidx.compose.animation.fadeOut
-import androidx.compose.animation.scaleIn
-import androidx.compose.animation.scaleOut
-import androidx.compose.animation.slideInHorizontally
-import androidx.compose.animation.slideOutHorizontally
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -36,49 +28,13 @@ import androidx.navigation.NavGraph.Companion.findStartDestination
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
-import com.example.seviya.core.designsystem.components.ClientBottomBar
 import com.example.seviya.core.designsystem.components.ClientTab
-import com.example.seviya.core.designsystem.components.FullScreenMenu
 import com.example.seviya.core.designsystem.components.GuestBottomBar
 import com.example.seviya.core.designsystem.components.GuestTab
-import com.example.seviya.core.designsystem.components.MenuOption
-import com.example.seviya.core.designsystem.components.WorkerBottomBar
 import com.example.seviya.core.designsystem.components.WorkerTab
-import com.example.seviya.core.navigation.CategoriesCatalog
-import com.example.seviya.core.navigation.ClientAgenda
-import com.example.seviya.core.navigation.ClientAlerts
-import com.example.seviya.core.navigation.ClientConfiguration
-import com.example.seviya.core.navigation.ClientDashboard
-import com.example.seviya.core.navigation.ClientFavorites
-import com.example.seviya.core.navigation.ClientMap
-import com.example.seviya.core.navigation.ClientMessages
-import com.example.seviya.core.navigation.ClientPaymentUpload
-import com.example.seviya.core.navigation.ClientProfile
-import com.example.seviya.core.navigation.ClientRequests
-import com.example.seviya.core.navigation.ClientSearch
-import com.example.seviya.core.navigation.ClientSettings
 import com.example.seviya.core.navigation.Landing
-import com.example.seviya.core.navigation.ProfessionalProfile
-import com.example.seviya.core.navigation.RequestAppointment
 import com.example.seviya.core.navigation.RoleAdmissionCatalog
 import com.example.seviya.core.navigation.RoleCatalog
-import com.example.seviya.core.navigation.TravelTimeConfig
-import com.example.seviya.core.navigation.WorkerAgenda
-import com.example.seviya.core.navigation.WorkerAlerts
-import com.example.seviya.core.navigation.WorkerAppointmentDetail
-import com.example.seviya.core.navigation.WorkerConfiguration
-import com.example.seviya.core.navigation.WorkerDashboard
-import com.example.seviya.core.navigation.WorkerMessages
-import com.example.seviya.core.navigation.ClientHome
-import com.example.seviya.core.navigation.WorkerPortfolio
-import com.example.seviya.core.navigation.WorkerProfile
-import com.example.seviya.core.navigation.WorkerReports
-import com.example.seviya.core.navigation.WorkerRequests
-import com.example.seviya.core.navigation.WorkerCategories
-import com.example.seviya.core.navigation.WorkerSchedule
-import com.example.seviya.core.navigation.WorkerServices
-import com.example.seviya.core.navigation.WorkerSettings
-import com.example.seviya.core.navigation.WorkersList
 import com.example.shared.presentation.requestAppointment.CurrentTimeSnapshot
 import com.example.shared.presentation.requestAppointment.RequestAppointmentDraft
 import com.example.seviya.core.designsystem.theme.AppTheme
@@ -89,11 +45,13 @@ import com.example.shared.domain.entity.AppointmentService
 import com.example.shared.domain.entity.Booking
 import com.example.shared.domain.entity.Service
 import com.example.shared.presentation.calendar.MonthlyCalendarViewModel
-import com.example.seviya.core.navigation.WorkerDailyAppointments
 import com.example.seviya.app.navigation.AppNavGraph
-import com.example.seviya.core.navigation.ClientLocationCatalog
-import compose.icons.TablerIcons
-import compose.icons.tablericons.*
+import com.example.seviya.feature.client.navigation.ClientFeatureBottomBar
+import com.example.seviya.feature.client.navigation.ClientFeatureMenu
+import com.example.seviya.feature.client.navigation.isClientScaffoldDestination
+import com.example.seviya.feature.worker.navigation.WorkerFeatureBottomBar
+import com.example.seviya.feature.worker.navigation.WorkerFeatureMenu
+import com.example.seviya.feature.worker.navigation.isWorkerScaffoldDestination
 import com.example.shared.utils.DateTimeUtils
 import org.koin.compose.viewmodel.koinViewModel
 
@@ -137,49 +95,11 @@ fun App() {
 
     val showClientBottomBar =
         sessionRole == SessionRole.CLIENT &&
-                (
-                        currentDestination.isRoute<ClientHome>() ||
-                                currentDestination.isRoute<CategoriesCatalog>() ||
-                                currentDestination.isRoute<WorkersList>() ||
-                                currentDestination.isRoute<ProfessionalProfile>() ||
-                                currentDestination.isRoute<ClientMap>() ||
-                                currentDestination.isRoute<ClientSearch>() ||
-                                currentDestination.isRoute<ClientAlerts>() ||
-                                currentDestination.isRoute<ClientDashboard>() ||
-                                currentDestination.isRoute<ClientAgenda>() ||
-                                currentDestination.isRoute<ClientProfile>() ||
-                                currentDestination.isRoute<ClientMessages>() ||
-                                currentDestination.isRoute<ClientConfiguration>() ||
-                                currentDestination.isRoute<ClientSettings>() ||
-                                currentDestination.isRoute<ClientFavorites>() ||
-                                currentDestination.isRoute<RequestAppointment>()||
-                                currentDestination.isRoute<ClientRequests>() ||
-                                currentDestination.isRoute<ClientPaymentUpload>() ||
-                                currentDestination.isRoute<ClientLocationCatalog>()
-                        )
+            currentDestination.isClientScaffoldDestination()
 
     val showWorkerBottomBar =
         sessionRole == SessionRole.WORKER &&
-                (
-                        currentDestination.isRoute<WorkerDashboard>() ||
-                                currentDestination.isRoute<WorkerAgenda>() ||
-                                currentDestination.isRoute<WorkerRequests>() ||
-                                currentDestination.isRoute<WorkerAlerts>() ||
-                                currentDestination.isRoute<WorkerProfile>() ||
-                                currentDestination.isRoute<WorkerMessages>() ||
-                                currentDestination.isRoute<WorkerConfiguration>() ||
-                                currentDestination.isRoute<WorkerSettings>() ||
-                                currentDestination.isRoute<WorkerReports>() ||
-                                currentDestination.isRoute<WorkerPortfolio>() ||
-                                currentDestination.isRoute<WorkerServices>() ||
-                                currentDestination.isRoute<WorkerSchedule>() ||
-
-                                currentDestination.isRoute<WorkerCategories>() ||
-                                currentDestination.isRoute<TravelTimeConfig>() ||
-                                currentDestination.isRoute<WorkerAppointmentDetail>() ||
-                                currentDestination.isRoute<WorkerDailyAppointments>()
-
-                        )
+            currentDestination.isWorkerScaffoldDestination()
 
     val guestCurrentTab =
         when {
@@ -221,72 +141,24 @@ fun App() {
                         }
 
                         showClientBottomBar -> {
-                            ClientBottomBar(
+                            ClientFeatureBottomBar(
+                                navController = navController,
                                 currentTab = currentClientTab,
-                                menuActive = clientMenuExpanded,
-                                onGoServices = {
-                                    clientMenuExpanded = false
-                                    workerMenuExpanded = false
-                                    currentClientTab = ClientTab.SERVICES
-                                    navController.navigateSingleTop(ClientHome)
-                                },
-                                onGoMap = {
-                                    clientMenuExpanded = false
-                                    workerMenuExpanded = false
-                                    currentClientTab = ClientTab.MAP
-                                    navController.navigateSingleTop(ClientMap)
-                                },
-                                onGoSearch = {
-                                    clientMenuExpanded = false
-                                    workerMenuExpanded = false
-                                    currentClientTab = ClientTab.SEARCH
-                                    navController.navigateSingleTop(ClientSearch)
-                                },
-                                onGoAlerts = {
-                                    clientMenuExpanded = false
-                                    workerMenuExpanded = false
-                                    currentClientTab = ClientTab.ALERTS
-                                    navController.navigateSingleTop(ClientAlerts)
-                                },
-                                onGoMenu = {
-                                    workerMenuExpanded = false
-                                    clientMenuExpanded = !clientMenuExpanded
-                                }
+                                menuExpanded = clientMenuExpanded,
+                                onCurrentTabChange = { currentClientTab = it },
+                                onClientMenuExpandedChange = { clientMenuExpanded = it },
+                                onWorkerMenuExpandedChange = { workerMenuExpanded = it }
                             )
                         }
 
                         showWorkerBottomBar -> {
-                            WorkerBottomBar(
+                            WorkerFeatureBottomBar(
+                                navController = navController,
                                 currentTab = currentWorkerTab,
-                                menuActive = workerMenuExpanded,
-                                onGoDashboard = {
-                                    clientMenuExpanded = false
-                                    workerMenuExpanded = false
-                                    currentWorkerTab = WorkerTab.DASHBOARD
-                                    navController.navigateSingleTop(WorkerDashboard)
-                                },
-                                onGoAgenda = {
-                                    clientMenuExpanded = false
-                                    workerMenuExpanded = false
-                                    currentWorkerTab = WorkerTab.AGENDA
-                                    navController.navigateSingleTop(WorkerAgenda)
-                                },
-                                onGoRequests = {
-                                    clientMenuExpanded = false
-                                    workerMenuExpanded = false
-                                    currentWorkerTab = WorkerTab.REQUESTS
-                                    navController.navigateSingleTop(WorkerRequests)
-                                },
-                                onGoAlerts = {
-                                    clientMenuExpanded = false
-                                    workerMenuExpanded = false
-                                    currentWorkerTab = WorkerTab.ALERTS
-                                    navController.navigateSingleTop(WorkerAlerts)
-                                },
-                                onGoMenu = {
-                                    clientMenuExpanded = false
-                                    workerMenuExpanded = !workerMenuExpanded
-                                }
+                                menuExpanded = workerMenuExpanded,
+                                onCurrentTabChange = { currentWorkerTab = it },
+                                onClientMenuExpandedChange = { clientMenuExpanded = it },
+                                onWorkerMenuExpandedChange = { workerMenuExpanded = it }
                             )
                         }
                     }
@@ -312,313 +184,40 @@ fun App() {
                 )
             }
 
-            AnimatedVisibility(
-                visible = clientMenuExpanded,
-                enter = slideInHorizontally(
-                    initialOffsetX = { it },
-                    animationSpec = tween(300)
-                ) + fadeIn(animationSpec = tween(220)) + scaleIn(
-                    initialScale = 0.98f,
-                    animationSpec = tween(300)
-                ),
-                exit = slideOutHorizontally(
-                    targetOffsetX = { it },
-                    animationSpec = tween(260)
-                ) + fadeOut(animationSpec = tween(180)) + scaleOut(
-                    targetScale = 0.98f,
-                    animationSpec = tween(260)
-                )
-            ) {
-                FullScreenMenu(
-                    title = "Menú cliente",
-                    options = clientMenuOptions(
-                        navController = navController,
-                        closeMenu = { clientMenuExpanded = false },
-                        onLogout = {
-                            clientMenuExpanded = false
-                            workerMenuExpanded = false
-                            sessionRole = SessionRole.GUEST
-                            currentClientTab = ClientTab.SERVICES
-                            requestAppointmentDraft = null
-                            navController.navigateToLandingClearingStack()
-                        }
-                    ),
-                    onDismiss = { clientMenuExpanded = false }
+            if (sessionRole == SessionRole.CLIENT) {
+                ClientFeatureMenu(
+                    navController = navController,
+                    menuExpanded = clientMenuExpanded,
+                    onClientMenuExpandedChange = { clientMenuExpanded = it },
+                    onLogout = {
+                        clientMenuExpanded = false
+                        workerMenuExpanded = false
+                        sessionRole = SessionRole.GUEST
+                        currentClientTab = ClientTab.SERVICES
+                        requestAppointmentDraft = null
+                        navController.navigateToLandingClearingStack()
+                    }
                 )
             }
 
-            AnimatedVisibility(
-                visible = workerMenuExpanded,
-                enter = slideInHorizontally(
-                    initialOffsetX = { it },
-                    animationSpec = tween(300)
-                ) + fadeIn(animationSpec = tween(220)) + scaleIn(
-                    initialScale = 0.98f,
-                    animationSpec = tween(300)
-                ),
-                exit = slideOutHorizontally(
-                    targetOffsetX = { it },
-                    animationSpec = tween(260)
-                ) + fadeOut(animationSpec = tween(180)) + scaleOut(
-                    targetScale = 0.98f,
-                    animationSpec = tween(260)
-                )
-            ) {
-                FullScreenMenu(
-                    title = "Menú trabajador",
-                    options = workerMenuOptions(
-                        navController = navController,
-                        currentWorkerId = currentWorkerId,
-                        closeMenu = { workerMenuExpanded = false },
-                        onLogout = {
-                            workerMenuExpanded = false
-                            clientMenuExpanded = false
-                            sessionRole = SessionRole.GUEST
-                            currentWorkerTab = WorkerTab.DASHBOARD
-                            requestAppointmentDraft = null
-                            navController.navigateToLandingClearingStack()
-                        }
-                    ),
-                    onDismiss = { workerMenuExpanded = false }
+            if (sessionRole == SessionRole.WORKER) {
+                WorkerFeatureMenu(
+                    navController = navController,
+                    currentWorkerId = currentWorkerId,
+                    menuExpanded = workerMenuExpanded,
+                    onWorkerMenuExpandedChange = { workerMenuExpanded = it },
+                    onLogout = {
+                        workerMenuExpanded = false
+                        clientMenuExpanded = false
+                        sessionRole = SessionRole.GUEST
+                        currentWorkerTab = WorkerTab.DASHBOARD
+                        requestAppointmentDraft = null
+                        navController.navigateToLandingClearingStack()
+                    }
                 )
             }
         }
     }
-}
-
-private fun clientMenuOptions(
-    navController: NavHostController,
-    closeMenu: () -> Unit,
-    onLogout: () -> Unit
-): List<MenuOption> {
-    return listOf(
-        MenuOption(
-            title = "Agenda",
-            subtitle = "Citas, historial y seguimiento",
-            icon = TablerIcons.CalendarEvent,
-            iconColor = Color(0xFF5FA8D3),
-            onClick = {
-                closeMenu()
-                navController.navigateSingleTop(ClientAgenda)
-            }
-        ),
-        MenuOption(
-            title = "Solicitudes",
-            subtitle = "Seguimiento de solicitudes y comprobantes",
-            icon = TablerIcons.Dashboard,
-            iconColor = Color(0xFF4F8CFF),
-            onClick = {
-                closeMenu()
-                navController.navigateSingleTop(ClientRequests)
-            }
-        ),
-        MenuOption(
-            title = "Favoritos",
-            subtitle = "Trabajadores guardados por el cliente",
-            icon = TablerIcons.Briefcase,
-            iconColor = Color(0xFFE77E9B),
-            onClick = {
-                closeMenu()
-                navController.navigateSingleTop(ClientFavorites)
-            }
-        ),
-        MenuOption(
-            title = "Mis Ubicaciones",
-            subtitle = "Gestiona tus direcciones frecuentes",
-            icon = TablerIcons.MapPin,
-            iconColor = Color(0xFF4A9EC7),
-            onClick = {
-                closeMenu()
-                navController.navigateSingleTop(ClientLocationCatalog)
-            }
-        ),
-        MenuOption(
-            title = "Perfil",
-            subtitle = "Datos personales e información de la cuenta",
-            icon = TablerIcons.User,
-            iconColor = Color(0xFF8E7CC3),
-            onClick = {
-                closeMenu()
-                navController.navigateSingleTop(ClientProfile)
-            }
-        ),
-        MenuOption(
-            title = "Mensajes",
-            subtitle = "Chats y conversaciones con trabajadores",
-            icon = TablerIcons.Message,
-            iconColor = Color(0xFF67B99A),
-            onClick = {
-                closeMenu()
-                navController.navigateSingleTop(ClientMessages)
-            }
-        ),
-        MenuOption(
-            title = "Dashboard",
-            subtitle = "Resumen general y actividad reciente",
-            icon = TablerIcons.Dashboard,
-            iconColor = Color(0xFFE29C7A),
-            onClick = {
-                closeMenu()
-                navController.navigateSingleTop(ClientDashboard)
-            }
-        ),
-        MenuOption(
-            title = "Configuración",
-            subtitle = "Opciones principales de la aplicación",
-            icon = TablerIcons.Settings,
-            iconColor = Color(0xFF9BB85D),
-            onClick = {
-                closeMenu()
-                navController.navigateSingleTop(ClientConfiguration)
-            }
-        ),
-        MenuOption(
-            title = "Ajustes",
-            subtitle = "Preferencias y personalización",
-            icon = TablerIcons.Adjustments,
-            iconColor = Color(0xFFD7B85A),
-            onClick = {
-                closeMenu()
-                navController.navigateSingleTop(ClientSettings)
-            }
-        ),
-        MenuOption(
-            title = "Cerrar sesión",
-            subtitle = "Salir de la cuenta y volver al inicio",
-            icon = TablerIcons.Logout,
-            iconColor = Color(0xFFEF4444),
-            onClick = onLogout
-        )
-    )
-}
-
-private fun workerMenuOptions(
-    navController: NavHostController,
-    currentWorkerId: String,
-    closeMenu: () -> Unit,
-    onLogout: () -> Unit
-): List<MenuOption> {
-    return listOf(
-        MenuOption(
-            title = "Perfil",
-            subtitle = "Datos personales e información pública",
-            icon = TablerIcons.User,
-            iconColor = Color(0xFF8E7CC3),
-            onClick = {
-                closeMenu()
-                navController.navigateSingleTop(WorkerProfile)
-            }
-        ),
-        MenuOption(
-            title = "Agenda Diaria",
-            subtitle = "Vista diaria en lista o mapa de las citas",
-            icon = TablerIcons.CalendarEvent,
-            iconColor = Color(0xFF4F8CFF),
-            onClick = {
-                closeMenu()
-                navController.navigateSingleTop(
-                    WorkerDailyAppointments(workerId = currentWorkerId)
-                )
-            }
-        ),
-        MenuOption(
-            title = "Servicios",
-            subtitle = "Registrar y modificar servicios ofrecidos",
-            icon = TablerIcons.Briefcase,
-            iconColor = Color(0xFF4CB5AE),
-            onClick = {
-                closeMenu()
-                navController.navigateSingleTop(WorkerServices)
-            }
-        ),
-        MenuOption(
-            title = "Mis Categorías",
-            subtitle = "Categorías de servicios que ofrezco",
-            icon = TablerIcons.Adjustments,
-            iconColor = Color(0xFF3B82F6),
-            onClick = {
-                closeMenu()
-                navController.navigateSingleTop(WorkerCategories)
-            }
-        ),
-        MenuOption(
-            title = "Tiempo de Traslado",
-            subtitle = "Tiempo de margen entre servicios",
-            icon = TablerIcons.Clock,
-            iconColor = Color(0xFFE2B100),
-            onClick = {
-                closeMenu()
-                navController.navigateSingleTop(TravelTimeConfig)
-            }
-        ),
-        MenuOption(
-            title = "Mensajes",
-            subtitle = "Chats y conversaciones con clientes",
-            icon = TablerIcons.Message,
-            iconColor = Color(0xFF67B99A),
-            onClick = {
-                closeMenu()
-                navController.navigateSingleTop(WorkerMessages)
-            }
-        ),
-        MenuOption(
-            title = "Configuración",
-            subtitle = "Opciones principales de la aplicación",
-            icon = TablerIcons.Settings,
-            iconColor = Color(0xFF9BB85D),
-            onClick = {
-                closeMenu()
-                navController.navigateSingleTop(WorkerConfiguration)
-            }
-        ),
-        MenuOption(
-            title = "Ajustes",
-            subtitle = "Preferencias y personalización",
-            icon = TablerIcons.Adjustments,
-            iconColor = Color(0xFFD7B85A),
-            onClick = {
-                closeMenu()
-                navController.navigateSingleTop(WorkerSettings)
-            }
-        ),
-        MenuOption(
-            title = "Reportes",
-            subtitle = "Indicadores, ingresos y desempeño",
-            icon = TablerIcons.ChartBar,
-            iconColor = Color(0xFF5F8EFA),
-            onClick = {
-                closeMenu()
-                navController.navigateSingleTop(WorkerReports)
-            }
-        ),
-        MenuOption(
-            title = "Portafolio",
-            subtitle = "Trabajos realizados y evidencia visual",
-            icon = TablerIcons.Photo,
-            iconColor = Color(0xFFC96AE6),
-            onClick = {
-                closeMenu()
-                navController.navigateSingleTop(WorkerPortfolio)
-            }
-        ),
-        MenuOption(
-            title = "Horario",
-            subtitle = "Días laborales, disponibilidad y zonas",
-            icon = TablerIcons.Clock,
-            iconColor = Color(0xFFE2B100),
-            onClick = {
-                closeMenu()
-                navController.navigateSingleTop(WorkerSchedule)
-            }
-        ),
-        MenuOption(
-            title = "Cerrar sesión",
-            subtitle = "Salir de la cuenta y volver al inicio",
-            icon = TablerIcons.Logout,
-            iconColor = Color(0xFFEF4444),
-            onClick = onLogout
-        )
-    )
 }
 
 @Composable
